Consolidate prior dataset collection wrappers, handled in item content now, disable generic wrappers
