--- client/src/components/History/ContentItem/DatasetCollection/index.js
@@ -1,6 +1,0 @@
-import DatasetCollection from "./DatasetCollection";
-
-export { default as DscUI } from "./DscUI";
-export { default as DscMenu } from "./DscMenu";
-export { default as JobStateProgress } from "./JobStateProgress";
-export default DatasetCollection;
