+++ client/src/components/History/ContentItem/GenericContentItem/GenericHistoryContent.vue
@@ -2,14 +2,13 @@
     <component :is="providerComponent" :id="data_item.id" v-slot="{ result: item, loading }">
         <div>
             <loading-span v-if="loading" message="Loading dataset" />
-            <component :is="renderComponent" v-if="item" :item="item"> </component>
+            <div v-if="item" :item="item"/>
         </div>
     </component>
 </template>
 
 <script>
 import { DatasetProvider, DatasetCollectionProvider } from "components/providers";
-import DatasetCollectionUIWrapper from "./DatasetCollectionUIWrapper";;
 import { STATES } from "components/History/model";
 import LoadingSpan from "components/LoadingSpan";
 
@@ -17,7 +16,6 @@ export default {
     components: {
         DatasetProvider,
         DatasetCollectionProvider,
-        DatasetCollectionUIWrapper,
         LoadingSpan,
     },
     props: {
@@ -27,10 +25,10 @@ export default {
     },
     computed: {
         renderComponent() {
-            return { hdca: "DatasetCollectionUIWrapper" }[this.data_item.src];
+            return null; // { hdca: "DatasetCollectionUIWrapper" }[this.data_item.src];
         },
         providerComponent() {
-            return { hda: "DatasetProvider", hdca: "DatasetCollectionProvider" }[this.data_item.src];
+            return null; // { hda: "DatasetProvider", hdca: "DatasetCollectionProvider" }[this.data_item.src];
         },
     },
     provide: {
