+++ client/src/components/History/CurrentCollection/Operations.vue
